Apply an accessible name to form.
The spec says that the "form" role only makes it a "landmark" if there
is an accessible name.

ARIA 1.3 says that the form role is a sub-role of landmark:
https://w3c.github.io/aria/, so if there is no accessible role then
the "form" role is inappropriate.

Since HTML-AAM doesn't say what the role should be otherwise (maybe
it should be "section"? That's what Chromium currently does), I
did not add a testcase for that situation.

Bug: 41490744

Change-Id: I6c6a93485a55a6645547cdf9e4eb4c8751e45141