Apply auction config's reportingTimeout.
Both top level and component auction config can specify a
reportingTimeout field to control the timeout of their reportWin() and
reportResult(). If component auction does not provide one, it will use
top level config's reporting timeout. A reporting timeout longer than 5
seconds will be clamped to 5s. If no reporting timeout is provided, a
default timeout kScriptTimeout is applied, as before this change.

Feature detection of this will be in a separate CL.

Fixed: b/326922003
Change-Id: Ibe2b41310dc44e6834a32ffa14afab0dd6ecb687